Moved the common test code
